+++ pkg/storages/splitmerge/splitmerge_test.go
@@ -1,4 +1,4 @@
-package storage
+package splitmerge
 
 import (
 	"bytes"
